requirements.txt
PyGithub
Networkx
tkintertable
matplotlib